requirements.txt
azure-cognitiveservices-speech==1.27.0
azure-core==1.26.4
azure-functions==1.13.3
azure-storage-blob==12.16.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
idna==3.4
importlib-metadata==6.6.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.11
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
